requirements.txt
streamlit
openai
requests
Pillow
python-dotenv
replicate
